backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

python-pptx==1.0.2
pillow==10.4.0
tqdm==4.66.5

pymupdf==1.24.9
pypdfium2==4.30.0

easyocr==1.7.2
torch==2.2.2
torchvision==0.17.2

numpy==1.26.4
imageio==2.37.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84

google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
protobuf==4.25.3

ipython==8.26.0
